% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{A brute-force 0/1 solution for the knapsack problem}
\arguments{
\item{x}{'x' is a dataframe with observations of value of item and weight of item.}

\item{W}{'W' is the maximum weight.}
}
\value{
Return the items that maximaze the value of the knapsack given the weight restraint, as well as the value of the knapsack.
}
\description{
The function calculates the combination of items to pack in a
knapsack to maximize the value in the knapsack given a weight restraint.
}
