% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{A greedy heuristics aproximation for the knapsack problem}
\arguments{
\item{x}{x' is a dataframe with observations of value of item and weight of item.}

\item{W}{'W' is the maximum wheight.}
}
\value{
The combination of items that maximizes the value packed given a
wheight restriction
}
\description{
The function calculates the combination of items to pack in a
knapsack to maximize the value in the knapsack.
}
